requirements_windows.txt
face_recognition_models>=0.3.0
face_recognition>=1.3.0
numpy
Pillow
scipy>=0.17.0
click>=6.0
opencv-python
dlib>=19.7

